Stage2_Labelling/requirements.txt
Package            Version
------------------ ------------
anyio              4.4.0
beautifulsoup4     4.12.3
certifi            2024.7.4
charset-normalizer 3.3.2
click              8.1.7
decorator          4.4.2
fastcore           1.7.0
fastlite           0.0.7
h11                0.14.0
httpcore           1.0.5
httptools          0.6.1
httpx              0.27.0
idna               3.7
imageio            2.35.0
imageio-ffmpeg     0.5.1
itsdangerous       2.2.0
moviepy            1.0.3
numpy              1.26.4
oauthlib           3.2.2
packaging          24.1
pillow             10.4.0
pip                24.0
proglog            0.1.10
python-dateutil    2.9.0.post0
python-dotenv      1.0.1
python-fasthtml    0.4.1
python-multipart   0.0.9
PyYAML             6.0.2
requests           2.32.3
setuptools         72.2.0
six                1.16.0
sniffio            1.3.1
soupsieve          2.6
sqlite_minutils    3.36.0.post4
starlette          0.38.2
tqdm               4.66.5
urllib3            2.2.2
uvicorn            0.30.6
uvloop             0.19.0
watchfiles         0.23.0
websockets         12.0
